netlify/functions/process_receipt/requirements.txt
google-generativeai==0.4.1
Flask==3.0.2
functions-framework==3.5.0
python-dotenv==1.0.1
